fix the crash when aliens start the beam
Had to work around this bug in oK interpreter:
https://github.com/JohnEarnest/ok/issues/96

basically, (a[,i]:,x) was being treated like a[i]:,x for scalars (i,x),
so i had to take the head of the index list when it contained only one
item. this meant that `sgn'` no longer worked, so i had to rewrite
sgn to work with both vectors and scalars.